Retry 8042 keyboard interrupts if host isn't responding
If the host somehow fails to see an edge on the keyboard IRQ line, it
won't read the 8042 data register.  The EC won't ever send another
IRQ, because it only does so after filling the register.  So the
keyboard will hang.

Work around this with a retry mechanism.  If the AP hasn't responded
after 3 additional keyboard events, generate another IRQ.  So a stuck
key will get unstuck if you tap it a few times.  That's reasonable,
and matches what people do already if they have a sticky key due to
crud accumulating in the keyboard.

I've tested this when the system is booted to the OS.  I don't see any
additional IRQs generated on the EC console ("KB extra IRQ"), so the
host is keeping up with the keyboard input stream.

If I'm in dev or recovery mode and bang on the keyboard right after
powering the system on (when the BIOS isn't yet paying attention to
the keyboard), I can see extra IRQs generated.  This shows the retry
mechanism is working.  The extra IRQs have no negative effect on the
boot process, and the keyboard works normally when the OS does
eventually boot.

BUG=chrome-os-partner:27222
BRANCH=rambi
TEST=Bang on the keyboard like a monkey.  Keyboard should still work.

Change-Id: Idd41b2d133267f48f959bca0cf062a18ca6551fb
Signed-off-by: Randall Spangler <rspangler@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/193272
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-by: Benson Leung <bleung@chromium.org>
Reviewed-by: Yung-chieh Lo <yjlou@chromium.org>